requirements.txt
scapy
pyserial
colorama
